+++ packages/components/src/base/error-boundary.tsx
