+++ packages/components/src/renderer/custom.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react'
+import ErrorBoundary from '../base/error-boundary'
 import { MessageTypeHandlerProps } from '../typings'
 import { pick } from '../utils'
-import ErrorBoundary from './Custom/ErrorBoundary'
-import Keyboard from './Keyboard'
+import Keyboard from './keyboard'
 
 const checkError = (moduleInjector: Function, component: string, payload: string): Error | null => {
   const errorPrepend = 'Custom component error: '
