+++ packages/components/src/renderer/dropdown.tsx
@@ -3,7 +3,7 @@ import Select from 'react-select'
 import Creatable from 'react-select/creatable'
 import { MessageTypeHandlerProps } from '../typings'
 import { renderUnsafeHTML } from '../utils'
-import Keyboard, { Prepend } from './Keyboard'
+import Keyboard, { Prepend } from './keyboard'
 
 export const Dropdown = ({
   options,
