+++ packages/components/src/renderer/index.tsx
@@ -2,13 +2,13 @@ import React, { ReactElement } from 'react'
 import { MessageType } from '../content-typings'
 import { Message, MessageTypeHandlerProps } from '../typings'
 import { Carousel, Card } from './carousel'
+import { QuickReplies } from './choice'
 import { Custom } from './custom'
 import { Dropdown } from './dropdown'
-import { Video, Audio, Image, File } from './FileMessage'
+import { Video, Audio, Image, File } from './file'
 import { LoginPrompt } from './login'
-import { QuickReplies } from './QuickReplies'
 import { Text } from './text'
-import { TypingIndicator } from './TypingIndicator'
+import { TypingIndicator } from './typing'
 import { VoiceMessage } from './voice'
 
 export const defaultTypesRenderers = {
