+++ packages/components/test/QuickReplies.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render } from '@testing-library/react'
 import React from 'react'
-import { Button } from '../src/renderer/QuickReplies/Button'
+import { Button } from '../src/base/button'
 
 describe('Quick Reply Button element', () => {
   test('it renders a simple button correctly', () => {
